pyproject.toml
[project]
name = "cp-autotests"
version = "0.1.0"
description = "Comprehensive test suite for Control Panel - Ethereum node deployment and management"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pytest>=7.4.3",
    "pytest-bdd>=7.0.1",
    "pytest-asyncio>=0.23.2",
    "pytest-xdist>=3.5.0",
    "pytest-timeout>=2.2.0",
    "pytest-rerunfailures>=13.0",
    "httpx>=0.25.2",
    "requests>=2.31.0",
    "schemathesis>=3.27.1",
    "playwright>=1.40.0",
    "pytest-playwright>=0.4.3",
    "testcontainers>=3.7.1",
    "factory-boy>=3.3.0",
    "faker>=21.0.0",
    "allure-pytest>=2.13.2",
    "locust>=2.20.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.3",
    "pydantic-settings>=2.1.0",
    "web3>=6.13.0",
    "eth-account>=0.10.0",
    "kubernetes>=28.1.0",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.2",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["config", "fixtures", "utils", "tests"]

[tool.uv]
dev-dependencies = [
    "ruff>=0.1.0",
    "black>=23.0.0",
    "mypy>=1.7.0",
]

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--alluredir=allure-results",
    "--clean-alluredir",
    "--browser=chromium",
    "--browser=firefox",
    "--browser=webkit",
]
markers = [
    "ui: UI tests using Playwright",
    "api: API tests",
    "bdd: BDD scenario tests",
    "core: Core functionality tests (ETH node deployment)",
    "rbac: RBAC and license tests",
    "workers: Worker and workflow tests",
    "nonfunctional: Performance, chaos, security tests",
    "smoke: Smoke tests",
    "critical: Critical path tests",
    "slow: Slow running tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "contract: Contract tests",
    "unit: Unit tests",
]
asyncio_mode = "auto"

[tool.pytest_bdd]
features_base_dir = "features/"

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.uv.scripts]
# Setup commands
setup = { cmd = ["sh", "-c", "uv sync && playwright install chromium firefox webkit && cp .env.example .env || true"] }
install = "uv sync"
install-browsers = "playwright install chromium firefox webkit"

# Test commands
test = "pytest -v"
test-ui = "pytest -v -m ui"
test-api = "pytest -v -m api"
test-smoke = "pytest -v -m smoke"
test-critical = "pytest -v -m critical"
test-core = "pytest -v -m core"
test-rbac = "pytest -v -m rbac"
test-workers = "pytest -v -m workers"
test-nonfunctional = "pytest -v -m nonfunctional"
test-parallel = "pytest -v -n auto"

# Reporting
report = { cmd = ["sh", "-c", "allure generate allure-results -o allure-report --clean && allure open allure-report"] }
report-generate = "allure generate allure-results -o allure-report --clean"
report-open = "allure open allure-report"

# Cleanup
clean = { cmd = ["sh", "-c", "rm -rf allure-results/ allure-report/ playwright-report/ test-results/ .pytest_cache/ && find . -type d -name '__pycache__' -exec rm -rf {} + 2>/dev/null || true && find . -type f -name '*.pyc' -delete"] }

# Linting and formatting
lint = "ruff check ."
lint-fix = "ruff check --fix ."
format = "black ."
format-check = "black --check ."
typecheck = "mypy ."

# Quality checks
check = { cmd = ["sh", "-c", "uv run ruff check . && uv run black --check . && uv run mypy ."] }

# Infrastructure
infra-up = "docker-compose up -d"
infra-down = "docker-compose down"
infra-logs = "docker-compose logs -f"
